Split and move health checker impl
This patch splits the health checker impl into two files: base and the
implementation of tcp, http and gRPC health checkers. The reason of
splitting is to let the health checker base to be extended by the
upcoming extension health checkers.
