+++ source/common/upstream/BUILD
@@ -93,48 +93,6 @@ envoy_cc_library(
 )
 
 envoy_cc_library(
-    name = "health_checker_lib",
-    srcs = ["health_checker_impl.cc"],
-    hdrs = ["health_checker_impl.h"],
-    external_deps = [
-        "grpc_health_proto",
-    ],
-    deps = [
-        ":host_utility_lib",
-        "//include/envoy/event:dispatcher_interface",
-        "//include/envoy/event:timer_interface",
-        "//include/envoy/grpc:status",
-        "//include/envoy/http:codec_interface",
-        "//include/envoy/http:codes_interface",
-        "//include/envoy/network:connection_interface",
-        "//include/envoy/network:filter_interface",
-        "//include/envoy/runtime:runtime_interface",
-        "//include/envoy/stats:stats_interface",
-        "//include/envoy/upstream:health_checker_interface",
-        "//source/common/buffer:buffer_lib",
-        "//source/common/buffer:zero_copy_input_stream_lib",
-        "//source/common/common:empty_string",
-        "//source/common/common:enum_to_int",
-        "//source/common/common:hex_lib",
-        "//source/common/common:logger_lib",
-        "//source/common/common:utility_lib",
-        "//source/common/grpc:codec_lib",
-        "//source/common/grpc:common_lib",
-        "//source/common/http:codec_client_lib",
-        "//source/common/http:header_map_lib",
-        "//source/common/http:headers_lib",
-        "//source/common/http:utility_lib",
-        "//source/common/network:filter_lib",
-        "//source/common/protobuf:utility_lib",
-        "//source/common/router:router_lib",
-        # TODO(mattklein123): The following should not be required once we allow pluggable health
-        # checkers.
-        "//source/extensions/filters/network/redis_proxy:conn_pool_lib",
-        "@envoy_api//envoy/api/v2/core:health_check_cc",
-    ],
-)
-
-envoy_cc_library(
     name = "host_utility_lib",
     srcs = ["host_utility.cc"],
     hdrs = ["host_utility.h"],
@@ -336,7 +294,6 @@ envoy_cc_library(
     srcs = ["upstream_impl.cc"],
     deps = [
         ":eds_lib",
-        ":health_checker_lib",
         ":logical_dns_cluster_lib",
         ":original_dst_cluster_lib",
         ":upstream_includes",
@@ -356,6 +313,7 @@ envoy_cc_library(
         "//source/common/network:utility_lib",
         "//source/common/protobuf",
         "//source/common/protobuf:utility_lib",
+        "//source/common/upstream/health_checker",
         "@envoy_api//envoy/api/v2/core:base_cc",
     ],
 )
