+++ test/common/upstream/BUILD
@@ -90,8 +90,8 @@ envoy_cc_test(
         "//source/common/json:json_loader_lib",
         "//source/common/network:utility_lib",
         "//source/common/protobuf:utility_lib",
-        "//source/common/upstream:health_checker_lib",
         "//source/common/upstream:upstream_lib",
+        "//source/common/upstream/health_checker",
         "//test/common/http:common_lib",
         "//test/extensions/filters/network/redis_proxy:redis_mocks",
         "//test/mocks/network:network_mocks",
